Week8/backend/requirements.txt

flask
flask-cors
pandas
numpy
faiss-cpu
sentence-transformers
openai
